Redesigned page 1, improved status check
