+++ BSMulti Installer2/Program.cs
@@ -16,7 +16,7 @@ namespace BSMulti_Installer2
         {
             Application.EnableVisualStyles();
             Application.SetCompatibleTextRenderingDefault(false);
-            Application.Run(new Form1());
+            Application.Run(new FindDirectory());
         }
     }
 }
